move article up a level without any nested routes
