+++ app/controllers/article.js
@@ -4,7 +4,7 @@ import Controller from '@ember/controller';
 import { inject } from '@ember/service';
 import { computed } from '@ember/object';
 
-import config from '../../config/environment';
+import config from '../config/environment';
 
 
 export default Controller.extend({
