+++ app/templates/article.hbs
@@ -1,1 +1,154 @@
-{{outlet}}
+<DoTargeting key='Template' value='Article' />
+{{remember-document-scroll key=model.id}}
+<div class="u-section-spacing l-container l-wrap">
+  <article class="c-article l-container u-spacing--double">
+    <NyprOArticleHeader as |header|>
+      <header.top as |top|>
+        {{#if model.breadcrumb}}
+          <top.breadcrumb @breadcrumbs={{model.breadcrumb}}/>
+        {{/if}}
+
+        <top.headline data-test-article-headline>
+          {{{model.title}}}
+        </top.headline>
+
+        <top.meta
+          @galleryPhotosCount={{model.gallery.slides.length}}
+          @galleryRoute='article.gallery'
+          @commentsUrl={{concat '#' commentsAnchor}}
+          @commentCount={{model.commentCount}}
+          data-test-article-meta
+          as |meta|>
+
+          <meta.byline @authors={{model.authors}} @sponsor={{model.sponsor}}/>
+          <meta.publishDate @publishDate={{model.publishedMoment}} />
+
+        </top.meta>
+
+      </header.top>
+
+      <header.lead as |lead|>
+        <lead.left>
+          <NyprMShareTools @vertical={{media 'isMediumAndUp'}} as |share|>
+            <share.share
+              @service='facebook'
+              @utm={{hash medium='social' source='facebook' campaign='shared_facebook'}}
+            />
+            <share.share
+              @service='twitter'
+              @params={{hash text=model.title via='gothamist'}}
+              @utm={{hash medium='social' source='twitter' campaign='shared_twitter'}}
+            />
+            <share.share
+              @service='reddit'
+              @params={{hash title=model.title}}
+              @utm={{hash medium='social' source='reddit' campaign='shared_twitter'}}
+            />
+            <share.share
+              @service='email'
+              @params={{hash body=(concat model.title ' - {{URL}}')}}
+              @utm={{hash medium='social' source='email' campaign='shared_email'}}
+            />
+          </NyprMShareTools>
+        </lead.left>
+
+        <lead.center>
+          {{#if model.hasGallery}}
+            <NyprMGalleryLead
+              @slides={{model.gallery.slides}}
+              @viewAll={{action 'viewGallery'}}
+              @goToSlide={{action 'goToSlide'}}
+            />
+          {{else if model.leadImage}}
+            {{#if model.leadImageLink}}
+              <a href={{model.leadImageLink}} target="_blank" rel="noopener">
+                <NyprMFigure
+                  class="c-lead"
+                  @srcS={{wagtail-image-url (hash id=model.leadImage.image) 640 null}}
+                  @caption={{model.leadImageCaption}}
+                  @credit={{model.leadImageCredit}}
+                />
+              </a>
+            {{else}}
+              <NyprMFigure
+                class="c-lead"
+                @srcS={{wagtail-image-url (hash id=model.leadImage.image) 640 null}}
+                @caption={{model.leadImageCaption}}
+                @credit={{model.leadImageCredit}}
+              />
+            {{/if}}
+          {{/if}}
+        </lead.center>
+
+        {{!-- <lead.right>
+          <NyprORelatedContent @items={{ARTICLE.relatedItems}}/>
+        </lead.right> --}}
+
+      </header.lead>
+
+    </NyprOArticleHeader>
+
+    <AdTagInside @slot="gothamist/interior/midpage/1" @contentsId={{model.moveableTypeId}}>
+      <NyprOArticleBody class={{unless model.hasLead 'u-space--zero--top'}} data-test-article-body>
+        <ArticleBody @blocks={{model.body}} />
+      </NyprOArticleBody>
+    </AdTagInside>
+
+    <div class="l-container--content u-section-spacing">
+      <NyprOArticleFooter as |footer|>
+        <footer.tags @tags={{model.displayTags}} as |Tag tag|>
+          <Tag @route={{array 'tags' tag}}/>
+        </footer.tags>
+
+        <footer.contact>
+          <p><strong>Do you know the scoop?</strong> &nbsp; {{#unless model.disableComments}}<a href="#{{commentsAnchor}}">Comment</a> below or {{/unless}}<a href="mailto:tips@gothamist.com" target="_blank" rel="noopener">Send us a Tip</a></p>
+        </footer.contact>
+
+        {{#if showTout}}
+          <footer.donate
+            @close={{action 'closeDonation'}}
+            @linkText='Donate Now'
+            @linkUrl='https://pledge3.wnyc.org/donate/gothamist/onestep?utm_medium=partnersite&utm_source=gothamist&utm_campaign=articledonationask'
+          >
+            <p>
+              Gothamist is non-profit local news. We rely on reader support. We need YOU so we can keep telling New York stories.
+            </p>
+          </footer.donate>
+        {{/if}}
+
+        <footer.ad>
+          <AdTagWide
+            @slot="gothamist/interior/midpage/2"
+            @breakMargins={{true}}
+            @showLabel={{true}}
+          />
+        </footer.ad>
+
+        {{#unless model.disableComments}}
+          <footer.comments>
+            <DisqusComments
+              @id={{commentsAnchor}}
+              @identifier={{model.moveableTypeId}}
+              @permalink={{model.permalink}}
+              @onReady={{navigateToComments}}
+            />
+          </footer.comments>
+        {{/unless}}
+      </NyprOArticleFooter>
+    </div>
+
+    <div class="l-container l-container--14col">
+      {{!-- <ArticleRecirc @article={{model}}/> --}}
+    </div>
+
+  </article>
+
+  <div class="c-article__read-more l-container l-container--14col">
+    <NyprMNavReadMore @items={{array
+      (hash route=(array 'sections' 'news') label='News')
+      (hash route=(array 'sections' 'arts-entertainment') label='Arts & Entertainment')
+      (hash route=(array 'sections' 'food') label='Food')
+      (hash route=(array 'popular') label='Popular')
+    }}/>
+  </div>
+</div>
